boards: Add refboard dummy board
Aim of this board is to provide a reference device tree that
will be used for dts/edts scripts sanity check.
While in theory this board does not use ARM architecture and should
have its own 'ref' arch, it is set under arm/ right now in order
to avoid providing a dedicated dummy arch.

Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>