Look up bindings via buses
Turn _compat2binding into a dictionary that maps 'compatible' strings to
dictionaries that map buses to bindings. This allows a single
'compatible' to have separate bindings depending on on which bus it
appears.

Represent bindings without buses with None for the bus key.

Use the parent bus (fetched from 'child: bus: ...' in the binding for
the parent) when looking up the bindings for nodes.