new_extract: Refactor and simplify alias generation
Path and instance aliases for registers work the same way. Only the
logic for the Device-derived part of the alias differs.

Move the logic for generating device aliases separately. That allows all
register aliases to use the same logic.

This is similar to the existing code (I think), but functions were
reduced down a bit further, to e.g. just return a list of Device
aliases.