samples: usb: cdc_acm*: Fix device testing
The device testing harness for the cdc_acm & cdc_acm_composite samples
has some expected console output to tell if the sample is running
propertly.  However that output that we validate against is handled via
LOG_DBG() and there might be delay in getting the output.  Switch the
output that we key on to printk instead of LOG_DBG().

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>