gen_defines: play tricks to benefit DT_INST users
Even though it was done for sound technical reasons, it's still
playing a bit of a fast one on Zephyr community to add access to all
device nodes at the last minute in the 2.3 release, including when
using DT_INST APIs.

In particular, instance numbers used to be allocated only to enabled
nodes, which is no longer true: now every node of a compatible is
allocated an instance number, even disabled ones.

This is especially unfortunate for drivers and applications that
expect singletons of their compatibles, and use DT_INST(0, ...) to
mean "the one enabled instance of my compatible".

To avoid gratuitous breakage, let's sort each edt.compat2nodes
sub-list so that enabled instances always come before disabled ones.

This doesn't break any API guarantees, because there basically *are*
no ordering guarantees, in part precisely to give us the flexibility
to do things like this. And it does help patterns that check
DT_NUM_INST_STATUS_OKAY(some_compat) and use instances 0 through
$VALUE-1, including the important singleton case.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>