ci: only rebase if required
Not all CI systems present the git checkout for a PR as merge of the
upstream repo + the PR (this is how shippable does it).  So detect if
the top commit is merge commit and only rebase if it is one.

This allows us to use the run_ci.sh script with other CI solutions like
github actions or buildkite that present the git checkout of a PR just
as a fetch of the PR head.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>