boards: arm: mps2-an521: Change SRAM location to PSRAM and resize flash
Currently SRAM region specified in dt for board mps2-an512 is set to
start at 0x3000_0000 and a 16M contiguous space is assumed. However, at
that address there is no such contiguous space of 16M, rather only a
128K area is available. As a consequence large applications linked with
Zephyr might end up using memory regions that are not valid, causing a
BusFault.

Application Note AN512 only specifies a 16M contiguous space available
starting at 0x8000_0000 (please see 'Table 3-4: SSRAM2 and SSRAM3
address mapping' and 'Table 3-6: External PSRAM mapping to Code Memory',
on pages 3-7 and 3-8, respectively), which resides in the PSRAM
(external RAM).

This commit sets the mps2-an521 board to use RAM starting at 0x8000_0000
so 16M of contiguous memory is in fact available as specified in the dt
and further used by the linker scripts responsible to generate the final
Zephyr ELF image.

Another issue with the current dt for that board is that it defines a
224M flash region which doesn't effectively exist, because most of it
is reserved (~148M). Hence that commit also modifies the dt to describe
the available memory regions used to run code (which are not exactly a
flash memory, but rather a SSRAM mapped into the code region that is
populated with code by auxilary hw when the board boots up).

This commit also updates the non-secure dts to reflect the changes in
the secure dts.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>