samples: subsys: ipc: Remove possible memory region overlap
On the MPS2 AN521 the non-secure configuration (mps2_an521_nonsecure)
serves double duty for the ipc tests that utilize a second core.  In
this config the RAM space for the is from 0x28100000 to 0x281fffff.  So
move the shared memory region right after that to 0x28200000 so its
doesn't possibly overlap in memory.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>