armclang: handle cases when build c++ code but file extension is .c
We have some tests cases like tests/subsys/logging/log_api that
build the same .c files as C or C++ for testing purposes.  In those
cases the testcases do something like:

  set_source_files_properties(main.c PROPERTIES LANGUAGE CXX)

Since the arm compiler is a single binary for C or C++ and normally
uses the file extension to determine C or C++ add an explicit
setting of -xc++ for when we are utilizing as a C++ compiler.  This
fixes any issues with filenames not being what the compiler expects.

Signed-off-by: Kumar Gala <kumar.gala@intel.com>